Make strlen benchmarks more deterministic
std::hash can still produce different results per run, but at least
now the glibc and aolc benchmarks should be using the exact same
strings as eachother (for any given length).
Doesn't matter supermuch for strlen, but does for strpbrk and such --
this is thus a good proof of concept.