Bandaid some issues with test_linkages
Tests were failing for strlen even when it was linking properly,
bc the given buffers were not memory-aligned.
Realistically I should just switch over to taking the addresses of
functions instead; that'll still die if they link improperly, but
it won't actually run the function or anything like that.
TODO later