Fix segfault issue with strlen
Previously, if the string began in the last 32B of a virtual-memory
page, we would read past the end of the page where the string started no
matter what; while this behavior wasn't caught by our tests previously,
the new strlen.Page test was designed with this in mind and did indeed
fail. Now, we drop back to the greatest 16B boundary at or below the
start of the string and align from there, then align onto a 128B
boundary before we begin the parallelized reads.